feat: larger matching pattern

+++ src/include/tlct/convert/helper/functional.hpp
@@ -43,7 +43,7 @@ namespace rgs = std::ranges;
     cv::Sobel(src, edges, CV_32F, 0, 1);
     edges = cv::abs(edges);
     intensity += (float)cv::sum(edges)[0];
-    intensity /= edges.size().area();
+    intensity /= (float)edges.size().area();
     return intensity;
 }
